rhi: metal: Fix incorrect native res. binding map check
Checking for nullptr is insufficient: just because there
is an empty map present, it does not mean it is valid. The
two cases must be handled identically.

This fixes a regression when using QShaders that do not
have an associated native resource binding map.

Amends 4639660dedceba7c16e1a8110bba16eff30be312

Change-Id: Icb239bf9a9261ed32f2cb7b22c60b608195618fc
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>