Sync behavior and documentation of QFoo::setItemDelegate()
QComboBox::setItemDelegate() does delete the old delegate which is in
constrast to all other setItemDelegate() functions which is quite
confusing. Sync in by *not* deleting the delegate in
QComboBox::setItemDelegate() and adjust the documentation to explicitly
state that the ownership is *not* passed to the affected classes.

[ChangeLog][QtWidgets][QComboBox] QComobBox::setItemDelegates no longer
deletes the previous delegate set.

Fixes: QTBUG-72483
Change-Id: I89c8e53903e7c9924a980c57b83ce40f5866e6ae
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>