Revert "Do not read Xft.dpi on platforms that shouldn't be using Xft settings"
This reverts commit c7fec68e1936576070d0fbac6cf40b818366d298.

This commit introduces a behavioural change within 5.14.

It's designed to special case plasma with a fix, but in practice it will
cause us more problems.

It will break:
 - font size on plasmashell and kwin on xcb which do not use Qt scaling
 - xwayland on projectors/headless tests

The original bug of double scaling that this was trying to fix is fixed
by b31852c4caa36cc564e25adbdacfa534e1dfe7c0 which is in 5.14.1 which
works in combination with the environment variables we set in plasma so
this is not needed.

Fixes: QTBUG-81532
Change-Id: I2f1b8ae4aecf7b80be4dbee812e6b4a64244fb1f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>