Move backing store OpenGL support to the platformcompositor module
QPlatformBackingStore had a dependency on the QOpenGLTextureBlitter, which is a
problem because we want to get rid of all the public QOpenGL* classes in the
gui module.

This splits the heavily QOpenGL dependent parts of the backing store
implementation into a separate class and moves it to the platformcompositor
module.

qplatformbackingstore.cpp is now mostly free from OpenGL implementation
details.

Platform integrations now have to explicitly request backing store OpenGL
support. This has been done for:

- xcb
- windows
- cocoa
- winrt
- android
- wasm
- ios

QPlatformGraphicsBufferHelper::lockAndBindToTexture is now exported so it can
be used from other modules.

Task-number: QTBUG-74409
Change-Id: I42ad9250e5a424939cf751a8ad880c7381ede2ae
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>