Make QMatrix4x4::projectedRotate public
The function is used by QGraphicsRotation, which is in widgets.
QGraphicsRotation is marked as a friend, so that it can access the
private method. QtQuick needs access to the same method, so internally
it declares a "fake" QGraphicsRotation class, just so that it can
private access. This breaks now that QtQuick needs to do QT += opengl,
which implies widgets and thus ends up including the header file that
also declares QGraphicsRotation.

So instead of these hacks, let's make the function public, but remain
\internal in terms of docs.

Change-Id: I671f68c9d7ab0aac985e0b583f63f3ee962567dc
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>